+++ Avalon/Demo/NetworkReadyState.cs
@@ -17,8 +17,8 @@ public class NetworkReadyState : State
 
     public Data Data { get; set; }
     public Range Range { get; set; }
-    private int Case { get; set; }
-    private int Status { get; set; }
+    private long Case { get; set; }
+    private long Status { get; set; }
 
     public override bool Execute()
     {
@@ -41,8 +41,8 @@ public class NetworkReadyState : State
 
         long count;
         count = 0;
-        
-        int cc;
+
+        long cc;
         cc = this.Case;
         if (cc == 0)
         {
